Move funapp and higher-order lowering into their own TUs
Cuts lower_funapp.cpp (resolve_builtin, fun_effectful, lower_funapp
and the density/eltwise/matrix/callable-transform/UDF families it
dispatches to) and lower_higher_order.cpp (the OdeSpec/Program
machinery: reduce_sum, quadrature, algebra, ODE and DAE lowering) out
of lower.cpp. Declarations were already in lower_internal.hpp from the
previous commit, so this is a pure cut-and-paste plus two new
CMakeLists.txt entries.

ctest: 120/120 passed.
dump_ops op-graph identity: 12/12 identical.
prep-trace census: diff -r census_before census_w2c_c2 -> IDENTICAL.
